# output/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Northwind Traders API",
    "description": "REST API for Northwind Traders - modernized from Access .accdb",
    "version": "v1"
  },
  "paths": {
...
    "/api/admin/settings/{id}": {
      "put": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SystemSettingDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemSettingDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemSettingDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/generate-orders": {
      "post": {
        "tags": [
          "Admin"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateOrdersDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateOrdersDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateOrdersDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GenerateOrdersDto": {
        "required": [
          "count"
        ],
        "type": "object",
        "properties": {
          "count": {
            "maximum": 100,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "SystemSettingDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "settingName": {
            "type": "string",
            "nullable": true
          },
          "settingValue": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateSettingDto": {
        "required": [
          "settingValue"
        ],
        "type": "object",
        "properties": {
          "settingValue": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      }
    }
  }
}