# output/src/ApiApp/Controllers/AdminController.cs
// Generated by booster-apidev | 2026-04-06 19:37:30

using ApiApp.Data;
using ApiApp.DTOs.Admin;
using ApiApp.Models;
using Microsoft.AspNetCore.Mvc;
using Microsoft.EntityFrameworkCore;

namespace ApiApp.Controllers;

[ApiController]
[Route("api/admin")]
public class AdminController : ControllerBase
{
    private readonly AppDbContext _db;

    public AdminController(AppDbContext db)
    {
        _db = db;
    }

    // GET /api/admin/settings
    [HttpGet("settings")]
    public async Task<ActionResult<IEnumerable<SystemSettingDto>>> GetSettings()
    {
        var settings = await _db.SystemSettings
            .Select(s => new SystemSettingDto
            {
                Id = s.Id,
                SettingName = s.SettingName,
                SettingValue = s.SettingValue,
                Notes = s.Notes
            })
            .ToListAsync();

        return Ok(settings);
    }

    // PUT /api/admin/settings/{id}
    [HttpPut("settings/{id:int}")]
    public async Task<ActionResult<SystemSettingDto>> UpdateSetting(int id, [FromBody] UpdateSettingDto dto)
    {
        var setting = await _db.SystemSettings.FindAsync(id);
        if (setting == null) return NotFound(new { error = $"Setting {id} not found." });

        setting.SettingValue = dto.SettingValue;
        await _db.SaveChangesAsync();

        return Ok(new SystemSettingDto
        {
            Id = setting.Id,
            SettingName = setting.SettingName,
            SettingValue = setting.SettingValue,
            Notes = setting.Notes
        });
    }

    // POST /api/admin/generate-orders
    [HttpPost("generate-orders")]
    public async Task<ActionResult> GenerateOrders([FromBody] GenerateOrdersDto dto)
    {
        // Find the internet sales employee by WindowsUserName
        var internetEmployee = await _db.Employees
            .FirstOrDefaultAsync(e => e.WindowsUserName == "internet-sales");

        if (internetEmployee == null)
        {
            return BadRequest(new { error = "Internet sales employee not found. Please seed the database first." });
        }

        var customerIds = await _db.Customers.Select(c => c.Id).ToListAsync();
        var products = await _db.Products.ToListAsync();

        if (customerIds.Count == 0)
            return BadRequest(new { error = "No customers found. Please seed the database first." });

        if (products.Count == 0)
            return BadRequest(new { error = "No products found. Please seed the database first." });

        var rng = new Random();
        var now = DateTime.UtcNow;

        for (int i = 0; i < dto.Count; i++)
        {
            var order = new Order
            {
                CustomerId = customerIds[rng.Next(customerIds.Count)],
                EmployeeId = internetEmployee.Id,
                OrderDate = now,
                Status = OrderStatus.New
            };
            _db.Orders.Add(order);
            await _db.SaveChangesAsync();

            int lineCount = rng.Next(2, 6);
            var usedProducts = new HashSet<int>();
            for (int j = 0; j < lineCount; j++)
            {
                var product = products[rng.Next(products.Count)];
                if (!usedProducts.Add(product.Id)) continue;

                _db.OrderDetails.Add(new OrderDetail
                {
                    OrderId = order.Id,
                    ProductId = product.Id,
                    Quantity = (short)rng.Next(5, 51),
                    UnitPrice = product.UnitPrice ?? 0m
                });
            }
            await _db.SaveChangesAsync();
        }

        return Ok(new { ordersCreated = dto.Count });
    }

    // DELETE /api/admin/test-data
    [HttpDelete("test-data")]
    public async Task<ActionResult> DeleteTestData()
    {
        var orderDetailsDeleted = await _db.OrderDetails.CountAsync();
        var ordersDeleted = await _db.Orders.CountAsync();
        var customersDeleted = await _db.Customers.CountAsync();
        var employeesDeleted = await _db.Employees.CountAsync();
        var productsDeleted = await _db.Products.CountAsync();

        // Delete in dependency order
        await _db.OrderDetails.ExecuteDeleteAsync();
        await _db.Orders.ExecuteDeleteAsync();
        await _db.Customers.ExecuteDeleteAsync();
        await _db.Employees.ExecuteDeleteAsync();
        await _db.Products.ExecuteDeleteAsync();

        return Ok(new
        {
            message = "Test data deleted.",
            customersDeleted,
            employeesDeleted,
            ordersDeleted,
            orderDetailsDeleted,
            productsDeleted
        });
    }
}