# output/src/ApiApp/DTOs/Admin/UpdateSettingDto.cs
// Generated by booster-apidev | 2026-04-06 19:37:30

using System.ComponentModel.DataAnnotations;

namespace ApiApp.DTOs.Admin;

public class UpdateSettingDto
{
    [Required]
    public string SettingValue { get; set; } = string.Empty;
}